Included fucnitonality for retretreiving links
